i735: Moved interim fields and calculation from AbstractRoutineAnalysis to AbstractAnalysis
